add go modules
