+++ kadai1/.gitignore
@@ -1,1 +1,2 @@
 _testdata
+_bin
