+++ kadai1/README.md
@@ -1,0 +1,20 @@
+課題1
+===
+
+## 画像変換コマンドを作ろう
+
+### 仕様
+- ディレクトリを指定する
+- 指定したディレクトリ以下のJPGファイルをPNGに変換
+- ディレクトリ以下は再帰的に処理する
+- 変換前と変換後の画像形式に指定できる(オプション)
+
+### 要件
+- mainパッケージと分離する
+- 自作パッケージと標準パッケージと準標準パッケージのみを使う
+- ユーザ定義型を使ってみる
+- GoDocを生成してみる
+
+
+
+
