+++ kadai1/cmd/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/gopherdojo/dojo6/kadai1/internal/cli"
+)
+
+func main() {
+	cli.Do()
+}
