+++ kadai1/go.mod
@@ -1,0 +1,12 @@
+module github.com/gopherdojo/dojo6/kadai1
+
+go 1.12
+
+require (
+	github.com/google/pprof v0.0.0-20190515194954-54271f7e092f // indirect
+	github.com/ianlancetaylor/demangle v0.0.0-20181102032728-5e5cf60278f6 // indirect
+	golang.org/x/arch v0.0.0-20190312162104-788fe5ffcd8c // indirect
+	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4 // indirect
+	golang.org/x/sys v0.0.0-20190712062909-fae7ac547cb7 // indirect
+	golang.org/x/tools v0.0.0-20190716021316-fefcef05abb1 // indirect
+)
